Objective expressions (#880)
* Added problem from Hull optimal control book

* Modified add_objective to accept expressions

* Added documentation

* Updated bibliography

* Added hull file to toc

* Fixed a bug in ODE evaluation group for shooting

* loosened tolerances on hull problem tests

+++ dymos/examples/hyper_sensitive/test/test_hyper_sensitive.py
@@ -16,7 +16,7 @@ from dymos import Trajectory, GaussLobatto, Phase, Radau
 from dymos.examples.hyper_sensitive.hyper_sensitive_ode import HyperSensitiveODE
 
 
-tf = np.float128(10)
+tf = 10
 
 
 @use_tempdirs
